requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0,<6
kaleido>=0.2.1,<1
streamlit>=1.30.0
pyyaml>=6.0
scipy>=1.11.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=7.4.0
openpyxl>=3.1.0
